20/12

+++ src/main/java/com/softdreams/excel/domain/Synthetic.java
@@ -129,7 +129,7 @@ public class Synthetic implements Serializable {
     private String caculationUnit;
 
     @Column(name = "amount")
-    private Long amount;
+    private Float amount;
 
     @Column(name = "price")
     private BigDecimal price;
@@ -426,16 +426,16 @@ public class Synthetic implements Serializable {
         this.caculationUnit = caculationUnit;
     }
 
-    public Long getAmount() {
+    public Float getAmount() {
         return this.amount;
     }
 
-    public Synthetic amount(Long amount) {
+    public Synthetic amount(Float amount) {
         this.setAmount(amount);
         return this;
     }
 
-    public void setAmount(Long amount) {
+    public void setAmount(Float amount) {
         this.amount = amount;
     }
